+++ components/server/src/routes/reports.py
@@ -10,7 +10,6 @@ from model.transformations import hide_credentials, summarize_report
 from server_utilities.functions import report_date_time
 
 
-@bottle.get("/api/v2/reports")
 @bottle.get("/api/v3/reports")
 def get_reports(database: Database):
     """Return the quality reports."""
@@ -26,7 +25,6 @@ def get_reports(database: Database):
     return overview
 
 
-@bottle.post("/api/v2/reports/attribute/<reports_attribute>")
 @bottle.post("/api/v3/reports/attribute/<reports_attribute>")
 def post_reports_attribute(reports_attribute: str, database: Database):
     """Set a reports overview attribute."""