+++ components/server/tests/routes/test_metric.py
@@ -3,7 +3,7 @@
 import unittest
 from unittest.mock import Mock, patch
 
-from routes.metric import delete_metric, get_metrics, post_metric_attribute, post_metric_new, post_metric_copy_v3, \
+from routes.metric import delete_metric, get_metrics, post_metric_attribute, post_metric_new, post_metric_copy, \
     post_move_metric
 
 from ..fixtures import JOHN, METRIC_ID, METRIC_ID2, REPORT_ID, REPORT_ID2, SOURCE_ID, SUBJECT_ID, SUBJECT_ID2, \
@@ -251,7 +251,7 @@ class MetricTest(unittest.TestCase):
 
     def test_copy_metric(self):
         """Test that a metric can be copied."""
-        self.assertEqual(dict(ok=True), post_metric_copy_v3(METRIC_ID, SUBJECT_ID, self.database))
+        self.assertEqual(dict(ok=True), post_metric_copy(METRIC_ID, SUBJECT_ID, self.database))
         self.database.reports.insert.assert_called_once()
         inserted_metrics = self.database.reports.insert.call_args[0][0]["subjects"][SUBJECT_ID]["metrics"]
         self.assertEqual(2, len(inserted_metrics))